% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Key_Pair.R
\name{key_Pair}
\alias{key_Pair}
\title{Key_Pair}
\usage{
key_Pair(df)
}
\arguments{
\item{df}{is expected to be a dataframe extracted from an SF object}
}
\value{
should return a dataframe of key value pairs
}
\description{
Takes a dataframe and converts it into a json key value pair currently only returns as strings
#'
}
\examples{
library(sf)
nc <- st_read( system.file("shape/nc.shp", package="sf"))

attri <- boundary \%>\% dplyr::mutate_all(~ gsub('"', "'", .)) \%>\%
as.data.frame() \%>\%  dplyr::select(-geometry) \%>\%  key_Pair()
}
