% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Polygon_Json_Creator.R
\name{Polygon_Json_Creator}
\alias{Polygon_Json_Creator}
\title{Polygon_Json_Creator}
\usage{
Polygon_Json_Creator(boundary, geometry = "geometry", wkid = "4326")
}
\arguments{
\item{boundary}{is expected to be a SF object with boundary data}

\item{geometry}{is expected to be a string with the field name containing the geometry data}

\item{wkid}{a text string indicating the well know id to be used.}
}
\value{
returns a json string
}
\description{
Takes an SF object with polygon data and coverts it to a ESRI format json string
}
\examples{

library(sf)
nc <- st_read( system.file("shape/nc.shp", package="sf"))

json <- esriformatr::Polygon_Json_Creator(nc)
}
